requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
alembic=1.4.1=pypi_0
blas=1.0=mkl
brotlipy=0.7.0=py37h2bbff1b_1003
ca-certificates=2021.7.5=haa95532_1
certifi=2021.5.30=py37haa95532_0
cffi=1.14.5=py37hcd4344a_0
chardet=4.0.0=py37haa95532_1003
click=8.0.1=pyhd3eb1b0_0
cloudpickle=1.6.0=pypi_0
colorama=0.4.4=pypi_0
cpuonly=1.0=0
cryptography=3.4.7=py37h71e12ea_0
cycler=0.10.0=py37_0
databricks-cli=0.14.3=pypi_0
dataclasses=0.8=pyh6d0b6a4_7
docker=5.0.0=pypi_0
entrypoints=0.3=pypi_0
filelock=3.0.12=pyhd3eb1b0_1
flask=2.0.1=pypi_0
freetype=2.10.4=hd328e21_0
gitdb=4.0.7=pypi_0
greenlet=1.1.0=pypi_0
huggingface_hub=0.0.13=py_0
icc_rt=2019.0.0=h0cc432a_1
icu=58.2=ha925a31_3
idna=2.10=pyhd3eb1b0_0
importlib-metadata=3.10.0=py37haa95532_0
importlib_metadata=3.10.0=hd3eb1b0_0
intel-openmp=2021.2.0=haa95532_616
itsdangerous=2.0.1=pypi_0
jinja2=3.0.1=pypi_0
joblib=1.0.1=pyhd3eb1b0_0
jpeg=9b=hb83a4c4_2
kiwisolver=1.3.1=py37hd77b12b_0
libpng=1.6.37=h2a8f88b_0
libprotobuf=3.14.0=h23ce68f_0
libtiff=4.2.0=hd0e1b90_0
libuv=1.40.0=he774522_0
lz4-c=1.9.3=h2bbff1b_0
mako=1.1.4=pypi_0
markupsafe=2.0.1=pypi_0
matplotlib=3.3.4=py37haa95532_0
matplotlib-base=3.3.4=py37h49ac443_0
mkl=2021.2.0=haa95532_296
mkl-service=2.3.0=py37h2bbff1b_1
mkl_fft=1.3.0=py37h277e83a_2
mkl_random=1.2.1=py37hf11a4ad_2
mlflow=1.18.0=pypi_0
networkx=2.6.1=pyhd3eb1b0_0
ninja=1.10.2=h6d14046_1
numpy=1.20.2=py37ha4e8547_0
numpy-base=1.20.2=py37hc2deb75_0
olefile=0.46=py37_0
openssl=1.1.1k=h2bbff1b_0
packaging=21.0=pyhd3eb1b0_0
pandas=1.2.5=py37hd77b12b_0
pillow=8.2.0=py37h4fa10fc_0
pip=21.1.3=py37haa95532_0
prometheus-client=0.11.0=pypi_0
prometheus-flask-exporter=0.18.2=pypi_0
protobuf=3.14.0=py37hd77b12b_1
pycparser=2.20=py_2
pyopenssl=20.0.1=pyhd3eb1b0_1
pyparsing=2.4.7=pyhd3eb1b0_0
pyqt=5.9.2=py37h6538335_2
pysocks=1.7.1=py37_1
python=3.7.10=h6244533_0
python-dateutil=2.8.1=pyhd3eb1b0_0
python-editor=1.0.4=pypi_0
python_abi=3.7=1_cp37m
pytorch=1.9.0=py3.7_cpu_0
pytz=2021.1=pyhd3eb1b0_0
pywin32=227=pypi_0
pyyaml=5.4.1=py37h2bbff1b_1
qt=5.9.7=vc14h73c81de_0
querystring-parser=1.2.4=pypi_0
regex=2021.7.6=py37h2bbff1b_0
requests=2.25.1=pyhd3eb1b0_0
sacremoses=master=py_0
scipy=1.6.2=py37h66253e8_1
seaborn=0.11.1=pyhd3eb1b0_0
setuptools=52.0.0=py37haa95532_0
sip=4.19.8=py37h6538335_0
six=1.16.0=pyhd3eb1b0_0
smmap=4.0.0=pypi_0
sqlalchemy=1.4.20=pypi_0
sqlite=3.36.0=h2bbff1b_0
sqlparse=0.4.1=pypi_0
tabulate=0.8.9=pypi_0
tk=8.6.10=he774522_0
tokenizers=0.10.3=py37_0
torchaudio=0.9.0=py37
torchvision=0.10.0=py37_cpu
tornado=6.1=py37h2bbff1b_0
tqdm=4.61.2=pyhd3eb1b0_1
transformers=4.8.1=py_0
typing-extensions=3.10.0.0=hd3eb1b0_0
typing_extensions=3.10.0.0=pyh06a4308_0
urllib3=1.26.6=pyhd3eb1b0_1
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
waitress=2.0.0=pypi_0
websocket-client=1.1.0=pypi_0
werkzeug=2.0.1=pypi_0
wheel=0.36.2=pyhd3eb1b0_0
win_inet_pton=1.1.0=py37haa95532_0
wincertstore=0.2=py37_0
xz=5.2.5=h62dcd97_0
yaml=0.2.5=he774522_0
zipp=3.5.0=pyhd3eb1b0_0
zlib=1.2.11=h62dcd97_4
zstd=1.4.9=h19a0ad4_0
